Rename testapp, addMavenDependency for Edge, update failed unit test
Rename testapp, addMavenDependency for Edge, update failed unit test
