+++ code/app/build.gradle.kts
@@ -5,10 +5,10 @@ plugins {
 }
 
 android {
-      namespace = "com.adobe.marketing.mobile.edge.consent.consentTestApp"
+      namespace = "com.adobe.marketing.mobile.edge.consent.testapp"
 
     defaultConfig {
-        applicationId = "com.adobe.marketing.mobile.edge.consent.consenttestapp"
+        applicationId = "com.adobe.marketing.mobile.edge.consent.testapp"
         compileSdk = BuildConstants.Versions.COMPILE_SDK_VERSION
         minSdk = BuildConstants.Versions.MIN_SDK_VERSION
         targetSdk = BuildConstants.Versions.TARGET_SDK_VERSION
