+++ code/app/src/main/java/com/adobe/marketing/mobile/edge/consent/testapp/ConsentTestApplication.java
@@ -9,7 +9,7 @@
   governing permissions and limitations under the License.
 */
 
-package com.adobe.marketing.mobile.edge.consent.consentTestApp;
+package com.adobe.marketing.mobile.edge.consent.testapp;
 
 import android.app.Application;
 import android.util.Log;
