+++ code/gradle.properties
@@ -14,5 +14,6 @@ mavenUploadDryRunFlag=false
 
 # production versions for production build
 mavenCoreVersion=3.0.0
+# TODO: CHANGE TO 3.0.0
 mavenEdgeVersion=2.0.0
 androidxAnnotationVersion=1.0.0
